Extract logic from cmd/config_edit into action.ConfigEdit
